+++ scripts/repo/winesapos-build-repo.sh
@@ -60,7 +60,7 @@ makepkg_fn linux-apfs-rw-dkms-git
 makepkg_fn linux-firmware-asus
 makepkg_fn linux-firmware-valve
 # Remove source packages downloaded by the 'linux-firmware-valve' PKGBUILD.
-rm -f ${OUTPUT_DIR}/linux-firmware-neptune*
+rm -f ${OUTPUT_DIR}/linux-firmware-neptune* ${OUTPUT_DIR}/steamdeck-dsp-*
 makepkg_fn macbook12-spi-driver-dkms
 makepkg_fn oxp-sensors-dkms-git
 makepkg_fn reiserfs-defrag